backend/requirements.txt
Django==4.2.26
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.0.1
Pillow==11.3.0
psycopg2-binary==2.9.10
channels==4.2.0
channels-redis==4.2.1
redis==5.2.1
